move chat logic to new file
