+++ assets/js/app.js
@@ -9,6 +9,6 @@ import "../css/app.scss"
 //
 // Import deps with the dep name or local files with a relative path, for example:
 //
-import socket from "./socket"
+import "./chat_room"
 //
 import "phoenix_html"
